Refactor: Profile Style 분리
