+++ src/components/Profile/ProfileBtn/ProfileBtn.jsx
@@ -1,42 +1,15 @@
 import axios from "axios";
 import React from "react";
 import { useNavigate } from "react-router-dom";
-import styled, { css } from "styled-components";
-import IconMessage from "../../assets/images/icon-message-circle.svg";
-import IconShare from "../../assets/images/icon-share.svg";
-import Button from "../common/Button/Button";
-import { ButtonStyle } from "../common/Button/Button";
-
-const FlexCommon = css`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-
-const InformationBottomSection = styled.section`
-  ${FlexCommon};
-  padding-bottom: 26px;
-  border-bottom: 1px solid #dbdbdb;
-`;
-
-const ImgCircleBtn = styled.button`
-  width: 34px;
-  height: 34px;
-  border: 1px solid #dbdbdb;
-  background-color: transparent;
-  border-radius: 50%;
-  box-sizing: border-box;
-  ${FlexCommon}
-`;
-
-const FollowBtn = styled(ButtonStyle)`
-  margin: 0 10px;
-`;
-
-const AddBtn = styled(ButtonStyle)`
-  width: 100px;
-  margin-left: 12px;
-`;
+import IconMessage from "../../../assets/images/icon-message-circle.svg";
+import IconShare from "../../../assets/images/icon-share.svg";
+import Button from "../../../components/common/Button/Button";
+import {
+  InformationBottomSection,
+  ImgCircleBtn,
+  FollowBtn,
+  AddBtn,
+} from "./ProfileBtnStyle";
 
 export default function ProfileBtn({
   type,
