+++ src/pages/Profile/Profile.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import styled from "styled-components";
-import ProfileInformation from "../../components/Profile/ProfileInformation";
+import ProfileInformation from "../../components/Profile/ProfileInformation/ProfileInformation";
 import PostList from "../../components/Post/PostList/PostList";
-import RecommendList from "../../components/Profile/RecommendList";
+import RecommendList from "../../components/Profile/RecommendList/RecommendList";
 import Modal from "../../components/Modal/Modal";
 import Navigation from "../../components/common/Nav/Navigation";
 import Alert from "../../components/Modal/Alert";
